Convert amount from float to int

+++ piggybank-rest/src/main/java/io/github/joaovicente/piggybank/CreateCreditRequestDto.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CreateCreditRequestDto {
     private String description;
-    private float amount;
+    private int amount;
 }
